refactor: 컨트롤러 의존성 분리
