+++ payment/payment-controller/src/main/java/shop/woowasap/payment/controller/PayController.java
@@ -13,9 +13,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import shop.woowasap.auth.domain.in.LoginUser;
-import shop.woowasap.order.domain.exception.DoesNotFindOrderException;
 import shop.woowasap.payment.controller.request.PayRequest;
 import shop.woowasap.payment.domain.PayType;
+import shop.woowasap.payment.domain.exception.DoesNotFindOrderException;
 import shop.woowasap.payment.domain.exception.DoesNotFindPaymentException;
 import shop.woowasap.payment.domain.exception.PayUserNotMatchException;
 import shop.woowasap.payment.domain.in.PaymentUseCase;
