+++ payment/payment-service/src/main/java/shop/woowasap/payment/service/PaymentService.java
@@ -7,11 +7,11 @@ import org.springframework.transaction.annotation.Transactional;
 import shop.woowasap.core.id.api.IdGenerator;
 import shop.woowasap.core.util.time.TimeUtil;
 import shop.woowasap.order.domain.Order;
-import shop.woowasap.order.domain.exception.DoesNotFindOrderException;
 import shop.woowasap.order.domain.in.OrderConnector;
 import shop.woowasap.order.domain.in.event.PaySuccessEvent;
 import shop.woowasap.payment.domain.PayStatus;
 import shop.woowasap.payment.domain.Payment;
+import shop.woowasap.payment.domain.exception.DoesNotFindOrderException;
 import shop.woowasap.payment.domain.exception.PayUserNotMatchException;
 import shop.woowasap.payment.domain.in.PaymentUseCase;
 import shop.woowasap.payment.domain.in.request.PaymentRequest;
